+++ Triviador/Question.h
@@ -5,7 +5,7 @@
 class Question
 {
 public:
-	Question(const std::string& question,const std::vector<std::string>& answears, const uint8_t& correctAnswear);
+	Question(const std::string& question,const std::vector<std::string>& answears, const int& correctAnswear);
 public:
 	std::string getQuestion();
 	std::vector<std::string> getAnswears();
